Update Gremlin.Net to 3.4.2 #16
Gremlin.Net now uses a connection pool. This pool can be configured
via the added `ConnectionPoolSettings` that are now also exposed in the
`JanusGraphClientBuilder`.
`new Graph().Traversal()` is now deprecated and the static method
`Traversal()` should be used instead.[1]
Since the `GremlinClient` now automatically creates a connection to the
server on construction instead of waiting for the first request, it was
necessary to change the way we check for the availability of the test
container.

[1]: http://tinkerpop.apache.org/docs/3.4.2/upgrade/#_anonymoustraversalsource

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>