Use official JanusGraph Docker image #19
The graph of the gods which we use for the tests is now created with a
groovy script, thanks to the new Docker image
(JanusGraph/janusgraph-docker#16). Having this test data already in the
image was the only reason to use a custom one.
This also updates test dependencies, especially to use newer
Testcontainers features.

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>