Clarify distribution licenses [skip ci]
The current statements conflates licenses under which user contributions are
accepted (which are inline in the bullet points), and the licenses under which
the project is distributed.

The new format clarifies that they are the same, and that the licenses apply
separately to code or documentation, and they apply equally whether we're
talking about a user contribution of such assets, or project distribution of
those same assets.

This is similar to the change being made in the JanusGraph FoundationDB storage
adapter in https://github.com/JanusGraph/janusgraph-foundationdb/pull/40

Signed-off-by: Misha Brukman <mbrukman@google.com>