Add basic GraphBinary support
Only RelationIdentifier and Text predicates are supported right now.
Geoshapes need to wait for JanusGraph 1.0.0 because of
JanusGraph/janusgraph#2856.

I also changed the API of the JanusGraphClientBuilder a bit to be more
independent of GraphSON and to be more in line with Gremlin.NET. It now
simply takes an `IMessageSerializer`.

GraphSON now uses JanusGraph predicates by default. So, users need to
change that on their own if they want to connect with a JanusGraph
Server < 0.6.0.
That should be OK however given that this contribution includes breaking
changes any way and it makes it easier to remove this functionality in
an upcoming version completely.

This also includes an update of Gremlin.Net to 3.5.2 as we need
TINKERPOP-2646[1] for this change.

Fixes #82

[1]: https://issues.apache.org/jira/browse/TINKERPOP-2646

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>