Deprecate JanusGraphClientBuilder
The reasoning behind this deprecation is mostly that the builder makes
it harder to configure the GremlinClient as it needs to replicate all
config options. Showing users how they can just configure a
GremlinClient with serialization support for JanusGraph also makes it
easier to use this library for existing users of Gremlin.Net.

I also fixed some outdated URLs.

Fixes #112

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>