0.4.2 Release
The main change in this release is the deprecation of the
JanusGraphClientBuilder.
The reasoning behind this deprecation is mostly that the builder makes
it harder to configure the GremlinClient as it needs to replicate all
config options. Showing users how they can just configure a
GremlinClient with serialization support for JanusGraph also makes it
easier to use this library for existing users of Gremlin.Net.

More information can be found in this issue:
https://github.com/JanusGraph/janusgraph-dotnet/issues/112

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>