Fix dotnet pack in workflow CTR
There was a breaking change in the .NET SDK that throws an error when
`dotnet pack` is used with the `-o` flag to specify an output directory
for a solution.
For more information, see: https://github.com/dotnet/sdk/issues/30624

We now instead have to tell `dotnet pack` to only pack the src project
itself and not the whole solution to avoid this.

I've also slightly improved build performance here by skipping
additional in `dotnet test` and `dotnet pack`.

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>