Initial Geoshapes support for GraphBinary
Only `Point` is supported for now, just like with GraphSON.
The implementation is very similar to the original one in Java with
exactly the same class structure:
`GeoshapeSerializer`: For general Geoshape serialization.
`GeoshapeTypeSerializer`: Base class for Geoshape type specific
serializers
`PointSerializer`: The first specific serializer for the `Point` type.

Fixes #127

Signed-off-by: Florian Hockmann <fh@florian-hockmann.de>